Add CookCLI python version
Implementation details:
* Terminal commands for:
  # listing files
  # reading recipe
  # adding recipes to shopping list, and printing the list
  # starting FLASK server

* Webpage displaying all recipes found undec @PWD/recipes
  that end in .cook.

Style-checking:
The code has been checked with black, flake8 and pylint.
Pylint returns some areas of improvement, but returns a score of above 9
which should be fine for now.

Adding mypy to this list would be nice.

Work to be done:
There is some work remaining. Mainly packaging of the project so
that it is installable on most systems. This could preferrable be
done via Docker of some sort to allow platform cross-compability.

The documentation could use an overhaul aswell.